MAVLINK_HIL_GPS: fixed CoG [0,360] angle wrap bug
The MAVLINK_HIL_GPS Course-over-Ground (cog) field
is a uint16_t expressed in hundredth's of degrees in
the range [0, 360).

The previous code was using the output of std::atan2
to populate the field, which outputs angles from [-pi, pi],
and it was not accounting for the necessary shift in
the interval to an unsigned representation.

The new code adds a helper function to common.h to
output a gazebo::math::Angle in degrees from 0-360, and
then makes use of the helper function when packing the
MAVLINK_HIL_GPS message in gazebo_mavlink_interface.cpp.