Changing positioning of HippoCampus and of the IMU
The position of the IMU relative to the HippoCampus has been changed in
order to get exact orientation data in NED frame from the control state topic.

The placement of the HippoCampus at the beginning of the Simulation has been changed
in oder to start at zero rotation in reagrd to NED frame.