Allow update rates of multiple of 250 Hz
It seems that some models like VTOLs or the camera gimbal require the
real_time_update_rate to be higher than 250 Hz, e.g. 500 Hz.

In this patch we loosen the hardcoded constraints of 250 Hz to multiples
of 250 Hz. We then skip the OnUpdate call to only interact with PX4 at
250 Hz.