Implement gimbal controller as MAVLink gimbal v2 device (#703)
* gazebo_gimbal_controller: add Mavlink interface

This is work in progress to add control the gimbal as a gimbal device
using the mavlink gimbal v2 protocol.

* gazebo_gimbal_controller: work on vehicle yaw

We now get the vehicle yaw from the mavlink message. This means we are
independent of the PX4 orientation and can also test the gimbal device
on its own using the gimbal_device_tester from the MAVSDK examples.

A nicer solution in the future might be to have a very simple state
estimator implemented in the controller which then uses the
autopilot_state_to_gimbal_device message for corrections.

This commit also implements yaw follow and lock mode.

* gazebo_gimbal_controller: send correct limits

* typhoon_h480: adapt gimbal limits

* gazebo_gimbal_controller: implement angle and rate

This should implement attitude control with gimbal angles,
angular rates, or both together.

* gazebo_gimbal_controller: switch UDP port

This is to connect to PX4.

* gazebo_gimbal_controller: fix follow mode angle

* gimbal_controller_plugin: fix limits

This was required after the mavlink headers were updated.

* geotagged_images: handle start/stop streaming cmds

This prevents unnecessary retransmissions on the QGC side.

* gimbal_controller_plugin: ignore unknown msgs

Let's not send unnecesssary acks. This otherwise triggers warnings such
as this in QGC:
VehicleLog: _handleCommandAck: mavlink component is ack'ing with
    different comp id:  154 instead of  100  for command 512

* gimbal_controller: add support for neutral mode

* models: remove inf from sdf as it's not valid

* Retune gimbal controller

Co-authored-by: Jaeyoung-Lim <jaeyoung@auterion.com>