Removed unused computation of CoG (#737)
* Removed unused computation of CoG
In the lift drag plugin there were a few lines of code used to compute the center of gravity of the respective link. Since the variables themselfes are unused I propose to remove them.

* Revert "Removed unused computation of CoG"

This reverts commit 97378945e26c16ce3341090c4b453731b2aeadd9.

Reverted changes to undo the formatting.

* Removed unused computation of CoG
In the lift drag plugin there were a few lines of code used to compute the center of gravity of the respective link. Since the variables themselfes are unused I propose to remove them.